+++ argilla-frontend/nuxt.config.ts
@@ -76,6 +76,7 @@ const config: NuxtConfig = {
     { src: "~/plugins/plugins/platform.ts" },
     { src: "~/plugins/plugins/language.ts" },
     { src: "~/plugins/plugins/color-schema" },
+    { src: "~/plugins/plugins/color-generator.ts" },
   ],
 
   // Auto import components (https://go.nuxtjs.dev/config-components)