+++ argilla-server/src/argilla_server/models/database.py
@@ -82,6 +82,10 @@ class Field(DatabaseModel):
     def is_image(self):
         return self.settings.get("type") == FieldType.image
 
+    @property
+    def is_chat(self):
+        return self.settings.get("type") == FieldType.chat
+
     def __repr__(self):
         return (
             f"Field(id={str(self.id)!r}, name={self.name!r}, required={self.required!r}, "
@@ -381,6 +385,11 @@ class Dataset(DatabaseModel):
     @property
     def distribution_strategy(self) -> DatasetDistributionStrategy:
         return DatasetDistributionStrategy(self.distribution["strategy"])
+
+    def field_by_name(self, name: str) -> Union["Field", None]:
+        for field in self.fields:
+            if field.name == name:
+                return field
 
     def metadata_property_by_name(self, name: str) -> Union["MetadataProperty", None]:
         for metadata_property in self.metadata_properties: