+++ argilla-server/tests/unit/api/handlers/v1/datasets/fields/test_list_dataset_fields.py
@@ -19,7 +19,7 @@ from httpx import AsyncClient
 
 from argilla_server.enums import FieldType
 
-from tests.factories import DatasetFactory, ImageFieldFactory
+from tests.factories import DatasetFactory, ImageFieldFactory, ChatFieldFactory
 
 
 @pytest.mark.asyncio
@@ -56,6 +56,41 @@ class TestListDatasetFields:
                     "dataset_id": str(dataset.id),
                     "inserted_at": image_field_b.inserted_at.isoformat(),
                     "updated_at": image_field_b.updated_at.isoformat(),
+                },
+            ]
+        }
+
+    async def test_list_dataset_fields_with_chat_field(self, async_client: AsyncClient, owner_auth_header: dict):
+        dataset = await DatasetFactory.create()
+        chat_field_a = await ChatFieldFactory.create(dataset=dataset)
+        chat_field_b = await ChatFieldFactory.create(
+            dataset=dataset, settings={"type": FieldType.chat, "use_markdown": False}
+        )
+
+        response = await async_client.get(self.url(dataset.id), headers=owner_auth_header)
+
+        assert response.status_code == 200
+        assert response.json() == {
+            "items": [
+                {
+                    "id": str(chat_field_a.id),
+                    "name": chat_field_a.name,
+                    "title": chat_field_a.title,
+                    "required": False,
+                    "settings": {"type": FieldType.chat, "use_markdown": True},
+                    "dataset_id": str(dataset.id),
+                    "inserted_at": chat_field_a.inserted_at.isoformat(),
+                    "updated_at": chat_field_a.updated_at.isoformat(),
+                },
+                {
+                    "id": str(chat_field_b.id),
+                    "name": chat_field_b.name,
+                    "title": chat_field_b.title,
+                    "required": False,
+                    "settings": {"type": FieldType.chat, "use_markdown": False},
+                    "dataset_id": str(dataset.id),
+                    "inserted_at": chat_field_b.inserted_at.isoformat(),
+                    "updated_at": chat_field_b.updated_at.isoformat(),
                 },
             ]
         }