+++ argilla-frontend/CHANGELOG.md
@@ -20,6 +20,8 @@ These are the section headers that we use:
 
 ## [2.1.0](https://github.com/argilla-io/argilla/compare/v2.0.1...v2.1.0)
 
+### Added
+
 - Added `DarkMode` ([#5412](https://github.com/argilla-io/argilla/pull/5412))
 - Added new `empty queue messages` ([#5403](https://github.com/argilla-io/argilla/pull/5403))
 - Added `HTML Sandbox` to support external and custom CSS and Javascript in fields ([#5353](https://github.com/argilla-io/argilla/pull/5353))